Add an Initialize method to RpcHandlerInterface
Currently there is no way to have initialization code that is run in a
handler object at the time it is registered via
Builder::RegisterHandler<T>. You can add a constructor, but that
constructor will not have access to the execution context, because
RpcHandlerInterface::SetExecutionContext<T>() is invoked *after* the
handler object is created.

This adds an Initialize() method to the interface, which is invoked
after the execution context has been set. This also has the nice benefit
of encouraging custom initialization in a method that isn't the
constructor, where exception safety is more of a concern.